requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
prometheus-client==0.9.0
python-json-logger==2.0.1
requests==2.25.1
urllib3==1.26.3
gitdb==4.0.7
GitPython==3.1.17
